refactoring webcli code
